packages/product-recommender/requirements.txt
pandas
numpy
scikit-learn
psycopg2-binary
pymysql
python-dotenv
ollama
